Refactor from "Visallo" to "OpenLumify"
Okay, so this technically *doesn't* need to be done, and in fact will make it more difficult to actually merge any upstream changes from Visallo, but it *did* force me to dive into the codebase and try to get a handle on where everything is and how it all hangs together.

There are of course still references to Visallo which I'd like to swap out, including logos and the like, and I will need to see if some Visallo-sponsored projects (such as Vertexium) can actually be used, but this was a fun little starting point.

Note that I've put things into a co.hackstreet group and org.openlumify package too; not sure I'll keep the co.hackstreet thing. Jus' playin'.